requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.9.0
grpcio==1.65.4
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langchain-openai==0.1.19
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langsmith==0.1.93
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.37.1
orjson==3.10.6
packaging==24.1
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pypdf==4.3.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
